+++ src/main/java/coordinates/SquareQuadrant.java
@@ -7,7 +7,7 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 //supports squares & rectangles
 public class SquareQuadrant extends AbstractQuadrant {
     @JsonCreator
-    public SquareQuadrant(@JsonProperty("xMul") int xMul, @JsonProperty("yMul") int yMul) {
+    public SquareQuadrant(@JsonProperty("xMul") double xMul, @JsonProperty("yMul") double yMul) {
         super(xMul, yMul);
     }
